Genetic data for the freshwater snail Drepanotrema depressissimum

Site-level data from Lamy et al. (2012).

dd.site

Format

A SpatialPointsDataFrame frame with 42 populations (combinations of site x year) and 22 variables, in the following slots:

coords

Spatial coordinates:
Longitude

West (WGS84)

Latitude

North (WGS84)

proj4string

Projection: "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0"

bbox

Boundary box: spatial extent

data

Site variables:
SiteID

Combination of SITE and YEAR

SITE

Site name (abbreviated)

YEAR

Year of sampling

Spatial

Logical: population included in spatial study?

MultiYear

Logical: population sampled in multiple years?

APE

Logical: population extinct between sampling years? Based on Table 5.

Cluster

Cluster name for populations that belong to a cluster.

n

Number of individuals sampled per population.

RA

Allelic richness based on a sample size of 14 individuals. From Table 2.

He

Gene diversity (expected heterozygosity). From Table 2.

f

Inbreeding coefficient. From Table 2

s

Estimate of selfing rate based on a multilocus maximum-likelihood method. From Table 2.

Type

Habitat type (p = pond, r = small river and s = swamp grasslands). From Table 4.

FST.GESTE

Population-specific FST as returned by GESTE. From Table 4.

Size

Patch size: Largest diameter in meter, tranformed with log(1 + X). From Table 4.

V

Vegetation cover: Fraction of site covered by aquatic vegetation (macrophytes and algae), tranformed with log(1 + X). From Table 4.

C

Connectivity to neighbouring sites, see paper. From Table 4.

Stab

Temporal stability, see paper. From Table 4.

D

Density of favourable habitats: Number of water bodies within a radius of 2 km. From Table 4.

APA

Apparent population age: Time in years since the most recent potential founder effect. From Table 4.

NLT

Long-term population size: Combination of observed density and site size averaged over years. From Table 4.

Fst.temp

Mean pairwise FST between temporal samples within site. From Table 5.

References

Lamy, T., Pointier, J. P., Jarne, P. and David, P. (2012), Testing metapopulation dynamics using genetic, demographic and ecological data. Molecular Ecology, 21: 1394–1410. doi:10.1111/j.1365-294X.2012.05478.x